Set Keystone endpoints to be v3 by default
This patch sets the admin, internal and public endpoints for
Keystone all to be v3 instead of the current mix of v2 for the
internal/public endpoints and v3 for the admin endpoint.

Existing deployments will have v3 endpoints added if they don't
already exist. The removal of v2 endpoints is left to the deployer
to do.

DocImpact
UpgradeImpact
Implements: blueprint liberty-release
Change-Id: I21b600b1bfb82edd9fba900ce6a9655f9addf9ed
Closes-Bug: #1477682