DEFCORE: Updated tempest config and resources
The change makes the tempest.conf discoverable apis use a comma
for separation. This is required so that the tests pass for
defcore otherwise they're skipped.

This change also adds in a second image within the default tempest
config such that it will now pass the defcore test to filter running
VMs by image type/name.

Change-Id: I8d6066ee88ef0e981c47fdc996db1421afb807c3
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>