Add telemetry testing to gate
This commit adds a new variable `tempest_service_available_aodh` to the
tempest role and adds a new function to openstack_tempest_gate.sh.j2 to
run telemetry tests.  We then update run-tempest.sh to also run the
telemetry tests.

At present we run all telemetry tests, but this is only ~ 9 tests which
complete very quickly.  In future we may want to be more limited to
what we run but it would be best to have someone with ceilometer
knowledge determine which those should be.

Change-Id: I8842e60c3c50f774119c8219f224069ff32394f8