Fix supplied identity uri
Change the var used to set [identity]/uri in tempest.conf so that it
uses the Keystone v2 API.

uri and uri_v3 need to be independently configurable. uri needs to be
set to the Keystone v2 API. This change is required for
openstack-ansible to pass the defcore Tempest tests. The test that
fails without this configurability is
tempest.api.identity.v2.test_tokens.TokensTest.

Change-Id: I190b2f397455f443fc24ea0031487a225b3222c3