Allocation pool for public subnet
Allow deployers to specify an allocation pool for the public subnet
created by tempest resources.

This is genreally useful, but the specific impetus is to avoid IP
conflicts between instances and the hosts on the public network in an
AIO. This patch does not solve the problem but paves the way for one that
will.

Informal-Dependency: I5e57bd43a2637c5e8508a7a0f0c6a41587aba94a
Change-Id: Ia58a6447ef28ba5d4ae2a2ac769ad0cc287f448d
Partial-Bug: #1564491