Fix broken installation task
Currently, when you try to install os_tempest in a functional test the
installation of tempest fails as the 'Install pip packages for tempest
(no prebuilt venv)' task is doing a with_items that includes
tempest_git_dest, which is incorrect.  This commit updates the
with_items in that specific task and also sets two common defaults in
the defaults/main.yml file.

Change-Id: I08510e2e15ed9bec08b0878f416a42a3b5b7ed6c