Remove volume-feature-enabled/snapshot override
The tempest.conf variable volume-feature-enabled/snapshot is set to
False by the tempest.conf.j2 template. This causes tests to be skipped
that would have passed if the configuration option had been set to True.

The commit removes the option from the template. This causes Tempest to
inherit the default value of True. To set this value back to False use
the var tempest_tempest_conf_overrides.

Change-Id: Ia7fe941c98f0f6bf820f24233e49f049f06667eb