Updates for tempest in functional tests
This commit does the following:

- adds tempest_log_dir variable so that /var/log/utility is not assumed
- adds oslo.concurrency and pyOpenSSL to tempest_pip_packages as
  tempest fails to run when it's built on the fly in a role's
  functional test

Note that a subsequent change has been made to openstack-ansible to set
tempest_log_dir to /var/log/utility.

Additionally, this role needs to be refactored so we can install
tempest similar to how other roles install their python software, which
would cause us to not have to track tempest dependencies in
tempest_pip_packages.

Depends-On: I907b7b416cf03d88c9b015b3265a216c51a0c7da

Change-Id: I74ee0f4df3066338f0810c322387399ba1a93a6b