Fix breaking typo and add functional test
Review [1] added a breaking typo to tasks/tempest_install.yml, which
we correct in this commit.

Additionally, this commit adds a functional test to os_tempest role
which deploys os_tempest itself and then runs some arbitrary keystone
tempest tests.  Having this test in place will prevent basic issues
like the one introduced in [1] from being commited.

Note that we deploy galera into a container, however we can probably
skip that.  Having it in a container gives us the ability to easily add
in more services if necessary.  The installation of rabbitmq has been
excluded since that should not be necessary in the context of this
testing.

[1] https://review.openstack.org/#/c/309034/

Change-Id: I144118b5e70f558554107345088cb36693a85e25