Standardise tempest role
All of our roles have a standard pattern for installing services,
however tempest does things slightly differently.  I think most of this
is legacy and this commit brings tempest in line (for the most part)
with how other roles are written.

Note that tempest now has a ``tempest`` cli, and in a future commit we
will update this role to use that instead.  This will allow us to skip
the additional step of cloning the tempest git repo to act as a working
directory when running tests.

For some background as to why this work is being done, newer versions
of tempest are failing when run in an IRR because we are not installing
urllib3 into tempest's venv.  This is not an issue when we do an
integrated build because the repo server actually installs tempest into
the tempest venv, and all requirements get satisfied.  However, when we
install tempest without a repo server we simply install
tempest_pip_packages (which previously did not include tempest itself)
into the tempest venv.  We could update update tempest_pip_packages to
include urllib3, but doing so creates duplicate work as this
requirement is already captured in tempest's requirements.txt.

Conflicts:
	defaults/main.yml
	tasks/tempest_install.yml
	tests/test-vars.yml

Change-Id: I6f3c36a8150b83afabae8d397d5fc7340dbc93fd
(cherry picked from commit 75bd092b9e66e4ef79bfc4449ec60901496128d4)