Ensure tempest runs independently in testing
In tempest role tests we install tempest on a container that already has
other services installed on it.

To ensure the tempest role is indepedent we should test it by deploying
it in it's own container.

We need to fix the dependencies for the tempest role by installing git
on the tempest host.

Change-Id: I0230f1a93d16243fc50df79b2d0780e744706c7b