Move tempest to Ansible 2.1.1 and central testing
Move tempest to use the central test repository as well as Ansible
2.1.1.

This mirrors the change made for the openstack-ansible-os_nova role,
whilst keeping tempest running on it's own container for testing
purposes.

Additionally this moves to use the testing scripts to simplify and make
the tox.ini file more uniform.

Co-Authored-By: Travis Truman <travis_truman@cable.comcast.com>
Change-Id: I06771dba022fa91bb0230a7c0ebe6d7872e80004