Install tempest plugins to correct venv
This change fixes the following bug:

The "Install tempest plugins (venv)" task attempts to install plugins
to the incorrect SHA-based venv instead of the venv created earlier
in the play by the "Create tempest venv dir" task. This results in
the playbook failing when it tries to use an old 8.0.x version of
pip.

Change-Id: I79edfb5adab863952c6fef966d32a6febee73366