Install test requirements for plugins
The ceilometer role is currently failing with the following
error:

ImportError: No module named gabbi

This is due to https://review.openstack.org/304500 adding a
new test which uses 'gabbi'. This exposes the fact that the
tempest plugin execution should have the dependencies in
test-requirements.txt installed into the tempest venv.

This patch ensures that the plugin installation also installs
all the test-requirements.

Change-Id: I9f2bece22a0098416f2d95f8a6003483e9821105