Reuse existing tempest workspace
If the tempest role is re-run against a host it will fail attempting to
re-initialize over an existing workspace. Update the changed_when and
failed_when conditions to allow reusing an existing workspace. Also
provide the path to the workspace rather than working from within the
tempest directory to avoid stderr and stdout being written to logs.

Change-Id: Ie88cb72b0d7cd3ef07ab269f2580485a7eeae3c6