Add cinder multi-backend tests
This commit updates templates/openstack_tempest_gate.sh.j2 and the
default tempest_test_whitelist to expose a cinder multi-backend test.
This is useful for ensuring that a) multiple backends are operational
(ceph/LVM for instance), while also testing that the multi-backend
features are functional.

NOTE: We also change the tests to use the sahara tempest plugin rather
      than designate as the designate plugin appears to be broken at
      the moment.

Change-Id: I4a730614cfb695b82a79dfbe540396e5f4fd198a