Do not make tempest execute by default
A deployer may want to deploy tempest, but not run it
immediately. For the tests repo and for the AIO we tell
it to run through vars/overrides, so we can set the role
default to not execute it instead.

Depends-On: I269b43200c0e8773b5c0c437f685f6a0a4cd4e5e
Change-Id: I8c6f841c71a22a43393d839cafb7d4584b9eb54d
(cherry picked from commit dc2ebe923e79cc7e0b8f7e2e70c6082cb06c3e09)