Add venv_tag local fact
In order to make it easier to detect the currently deployed
venv for a service, and therefore allow smarter decisions
for things like upgrading, we implement the venv tag as a
local fact.

The file used to store facts will be the same for all
OpenStack services, with each service using its own section.

Example:

"ansible_local": {
    "openstack_ansible": {
        "tempest": {
            "venv_tag": "14.2.1"
        }
    }
}

Change-Id: I79721f448badf36fd6071602f55254141d4d25f4