Add block/always to tempest result file tasks
The tempest result files should always be generated, regardless of
a tempest test failure or not. With out the block/always statement,
the tempest result files would only generate if the tempest tests
succeeded. This commit adds the block/always around the result file
tasks so that they are generated when the tempest tests fail.

Change-Id: I6698d52a7209005449fdb69a35eeaad763c59da9