update for changes needed to pass refstack
Without the image_ref_alt and subnet_allocation fix we pass 98.7%
https://refstack.openstack.org/#/results/218a2d45-e0c7-461a-957e-e5ca4c1271ad

With the following commit we pass 100%
https://refstack.openstack.org/#/results/2d2e8eaa-4cf5-4494-a418-97dbb1fee74a

A second image (image_ref_alt) needs to be uploaded for tests run by
tempest.  subnet_allocation also needs to be added to the list of
plugins in neutron's api_extensions within network-feature-enabled.

This gets us passing refstack in pike (not yet tested for queens or
rocky).

Change-Id: Ia31a8a9c5bac5b3bdbeded25ad9a9f3db2e62007