Install plugins for standard services
Provide a list of standard tempest plugins in variables, and
enable some of them by default. The ones enabled by default
are the ones set by default for the rest of the tempest role.

We are using --isolated for now, to ensure the installation
of the package can fetch the packages from anywhere, not
only from the repo server, because the repo server doesn't
build the tempest plugins requirements (we don't list them
in our tempest packages to install, because we find them
at runtime, instead of building them staticly in our code)

Change-Id: I4d504e9c60b6d15fa5a5fdf013c4392a9e895ac0