Switch from testrepository to stestr
This commit replaces the call to testr with stestr, and updates the
output files accordingly. This is necessary because tempest has moved
away from testrepository [1], which is resulting in empty files being
created.

Lastly, we update the functional tests to ensure the subunit results
files exist and are not empty.

[1] https://review.openstack.org/#/c/504345/

Depends-On: https://review.openstack.org/#/c/557988/
Change-Id: I2496fccff1f51c586e999538232a4bcf0addbee5
(cherry picked from commit f8404f5eeb2a07382782d63e126fae54b9383439)