Switch from testrepository to stestr
This commit replaces the call to testr with stestr, and updates the
output files accordingly. This is necessary because tempest has moved
away from testrepository [1], which is resulting in empty files being
created.

Lastly, we update the functional tests to ensure the subunit results
files exist and are not empty.

[1] https://review.openstack.org/#/c/504345/

Change-Id: I2496fccff1f51c586e999538232a4bcf0addbee5