Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: Icf39a1b5057e998d92923026a587ee3fdcb37c13
Partial-Bug: #1764470
(cherry picked from commit ce9d7ae0b31833037d21c79cf61d60970ba35bcb)