Prefer distribution packages for Tempest dependencies
Distributions provide packages for all the Tempest role dependencies
so we can use these instead of the ones from PIP in order to keep
the host as clean as possible. openSUSE provides suitable versions for
all of them, whereas Ubuntu provides everything except for shade which
we still need to get from PIP. Similarly, CentOS provides everything
except from virtualenv but this is being installed in the
openstack_hosts role.

Since we are still mixing distribution and pip packages, we need to
stop using upper-constrains when installing pip packages on the host
so distro packages can be used to satisfy their dependencies.

Change-Id: I5a78e2120e596d36629b4ba978b2b5df76b149b0
Implements: blueprint openstack-distribution-packages