Comment tempest_git_repo role default
This commit comments out the default for tempest_git_repo so that we
can install tempest from PyPi, but with constraints applied. This was
attempted in [1] but wasn't enough to achieve this outcome. Note that
this change also requires an update to the repo_build role [2] for it
to fully work.

[1] I014201129b4d6b47ec45ff6239287623123a1595
[2] I4468092eb5c86d6346a793ba1e2bfc6e772de3bb

Change-Id: I05f88153e522b55e619b4d2e2cdfd311ab8228d8