Pin cmd2 to the last python2 version of the library
The cmd2-0.9.0 is python3-only and the tempest venv is built
unconstrained, so it's trying to bring in the latest version
which is incompatible with python2. We pin the library to
ensure that the venv build completes correctly.

Depends-On: I35e3688ae632d5f7747f32dafc41d23883c033c5
Change-Id: Id5b9274ce39dec3d5ee295d7ef941bf42ca7cbd2
Signed-off-by: Maksim Malchuk <maksim.malchuk@gmail.com>
(cherry picked from commit 8c7b78424f699f19ea0a11e4c9dfee78cee1976f)