Remove unnecessary distro package installs
There appears to be no need for the openstack packages
to be installed by this role, especially given that
they conflict with the installs done by the utility
playbook when doing a source-based install.

Also, the other packages in the list appear to be a
remnant of the past. It appears that neither 'which'
or 'rsync' are used in the role through shell commands
or ansible modules.

Change-Id: Ib757dd2616f6d6aa6ebc6b870b76b3fa2ac7c8f6
(cherry picked from commit 7dcb2c7f008509967feed49ff0951c3b25c03b50)