Enable python-tempestconf support
This is an initial support for tempestconf tool.
Initially, tempestconf tool is being executed to generate the
tempestconf as it is, furthermore will be added the option to override
values, generate resources, hardcoded options, etc.
This is will be done in an incremental way in order to not have big
patches full of functionalities that might not be able to be fully
tested in one single patch.

Change-Id: I08cfa29361804de3bdb9077a7d8d2b8b417e50d8