Enable overriding the service setup host python interpreter
In order to enable the service setup host python interpreter to
be changed easily, we make it a variable. This will be useful
when someone sets the service setup host to be the utility
container, because we'll be able to set this var by default.

Change-Id: Ie3b2c2e04c25737653bd53890fcd847d4b3a430e
(cherry picked from commit 605a92baf77e275cecd7ab03a59375af940e8f5d)