Better tempest blacklist management
This patch set the information regarding how the tempest_test_blacklist
variable should be configured. From now, each test in the blacklist
should have a launchpad and/or bugzilla tighted on it in order to
trackdown the progress of the test, as well as use other tools to
automate the verification of the status of those bugs.

Change-Id: I36c41d97ca778d892c1655731cca9ac0112851c1