Added support for installing tempestconf from distro
On Red Hat, python-tempestconf is a dependency of
openstack-tempest and on SUSE, it is a seperate package and can
be installed on SUSE distro job only when tempest_use_tempestconf
is set to true there. On Ubuntu, we are running tempest from
source, so we donot need to make any change there.

Depends-On: https://review.openstack.org/#/c/625545/

Change-Id: I1644685a3aa263f74eada36b67f9d624e6393ddd