Automatically select the correct tempest plugins
In this patch we change the way the plugin list is used
by default to ensure that only the plugins needed for the
environment are installed.

This removes the need to implement override plugin lists
when testing other services, allowing us to centralise
all configuration and source-based build testing in this
role.

We also add any missing plugins or additional boolean
enablement variables to ensure we have a complete matching
list between the defaults and vars.

Note that a change to the actual tests exercised will come
in a later patch.

Change-Id: I576386e158604bc21abe3672df407fcf08684d3d