Add an ip address to eth12 in OSA test containers
The tempest tests need to ssh to booted virtual machines. The neutron
public network is 10.1.3.0/24, which is the flat network on br-vlan.
From the container perspective this was previously reached via the
default gateway on eth0 as there was no specific route to the
10.1.3.0/24 network.

On centos traffic is no longer being forwarded to the br-vlan flat network
via the container eth0 and host networking, but this still works for
other distributions.

This patch adds an IP address to the present, but previously un-addressed
interface eth12 in each container. This gives a direct route to the neutron
public network from the container and does not require the host network
to forward the traffic. The test setup is now more decoupled from the
host networking which should increase reliability.

Change-Id: Id8fbe05e83bc6d86f57202301e712fc4e8a17003