Ensure stackviz installs from tarball URL only
Stackviz installs from a tarball in order to make it happen quicker.
If we were to build it from git sources, then we'd have to build the
HTML page ourselves, which takes approximately ten minutes.

Due to the tarball URL being used as the package name, we're unable
to constrain it without making the venv build role a lot more complex.

To keep it simple, we just disable the wheel build process and install
it from the tarball instead.

Depends-On: https://review.openstack.org/644391
Change-Id: Ieb498c49377d77577c98ddb2d911066e44d8d9dd