Do ping test from tempest host and disable it by default
Currently the ping test runs from the tempest_service_setup_host,
which may not have the correct package installed to execute it.
The ping package is installed on to the tempest host, and the
requirement to do the ping test is actually to verify connectivity
from the tempest host before running tempest itself.

This patch changes the ping test to run from the tempest host,
and because the test is primarily for gate testing, it is disabled
by default.

Change-Id: I3d1f9e17ef45b4427e8d3088bbcfa0c0df126c4a