Update role for new source build process
The variable tempest_developer_mode and tempest_venv_download no
longer carry any meaning. This review changes os_tempest to do
the equivalent of what developer_mode was all the time, meaning
that it always builds the venv and never requires the repo
server, but it will use a repo server when available.

As part of this, we move the source build out of its own file
because it's now a single task to include the venv build role.
This is just to make it easier to follow the code.

Change-Id: I7ab81ab707a5576682abcbe9952d85e79a1f3856