Test using integrated build
In order to do a more complete verification of any patches,
we add a full uncontainerised OpenStack deployment to the
functional testing using the integrated repo. This replaces
the previous role functional test mechanism.

Any additional role tests are left as-is. They will require
some extra implementation in the integrated build before they
can be transferred.

Depends-On: https://review.openstack.org/647840
Depends-On: https://review.openstack.org/648502
Depends-On: https://review.openstack.org/648551
Change-Id: Idab108649076b24d526e9dd7d6569f306e220df2