Do not ping router when public net is local
Don't bother trying to ping the router address when the public
network type is 'local'. In calico, a local network type is used,
and the router is not set up on the host until an instance using
that network is created. Since this check occurs before the
tempest run (and before any instances are created), the gateway
is not expected to ping.

Change-Id: I47543778487b6a9a2972b29d270d199a4d1e9b61
(cherry picked from commit a1a5383b2c02fff3f43e6c47ed01f6b9a6576fce)