Temporarily set bionic job to non-voting
Due to the shift from git.openstack.org to opendev.org none of
the depends-on features are working. The patches to fix that
cannot merge until https://review.opendev.org/657778 merges.

In order to enable this to merge, we temporarily disable voting
for this failing job. Once the above match merges, we can revert
this patch.

Depends-On: https://review.opendev.org/657816
Change-Id: I75f116fc226fc1821ab88835c89b2e64a28c7dbf