Append the tempest run command output in file
In CI, when we run full tempest api and scenario tests, it takes
time and in most of the time at Execute tempest tasks hits
Timeout and no ara report got generated. So we have no idea
what tests ate the time.

Logging the tempest result in a file will give more info
on what happened with the tests.

Change-Id: Ibba761551e6bc14c127d9461de2a72e9f0fcb715
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>