Don't download image when it's already there
As image should have been already cached in CI [1], we shouldn't
download it again but just verify the checksum.

NOTE - at the time of this commit download.cirros.com is not working
so the cirros image is also switched to 0.4.0 which is present in
the infra CI image.

[1] https://opendev.org/openstack/openstack-ansible/src/branch/master/tests/roles/bootstrap-host/templates/user_variables.aio.yml.j2#L226

(cherry-picked from commit ad4841f70a0ad4ed8090fc3f4c83812204cb59c0)
Change-Id: Ib764c70e483513bec5bd97d4499e67beb4501826