post_install: add missing onditional to workspace task
The "Move over workspace when config is updated" task was running if
copy_tempest_config changed, but copy_tempest_config is created in
"Copy tempest config" only if "not tempest_use_tempestconf" conditional
is met. To avoid an error like "'copy_tempest_config' is undefined", we
need to add the same conditional to the first mentioned task.

Change-Id: I0d603dfa7319d09201f5f1d64d10cc71c56c2c21