Allow deployer to skip default resource creation
This patch is part of a set that aims to break out tempest
resource creation from the role, and introduces the
tempest_default_role_resources var. When true, default role
resources are created. When false, deployer-specific resources
must be created prior to executing os-tempest-install.yml or
including the os_tempest role in other playbooks.

Change-Id: I42cb8ffbe8aa8cfdcdcfcbbccf28117a27c519b1