defaults: use a versioned path for stackviz/tempestconf virtualenv
When upgrading to Ussuri from Train we already had a virtualenv
called 'stackviz'. As a result the upgrade process attempts to
re-use the same virtualenv, which contains a pip executable for
Python 2. When trying to install Python 3 packages this fails.

We worked around this by using the same path versioning mechanism
used for tempest and other components, forcing the creation of a
new virtualenv. This patch applies this approach to the
virtualenvs for stackviz and tempestconf.

Change-Id: Iad382dfeb91d25c325b41d3adc7f703e0e4688a8
(cherry picked from commit 4a40fc8b80274e6d07b55b52c494af251a5efe66)