Use network UUID for floating_network_name
``floating_network_name`` works well with UUIDs as well.

This change doesn't make any difference when tempest creates public
network.

But when it's not, operator does not set ``tempest_public_net_name``
which leaves ``floating_network_name`` with a defeault value.

In order to fix ``floating_network_name`` behavior for both scenarios,
it should be set to the ``tempest_neutron_public_network_id``.

We do the same thing for ``public_network_id``.

Change-Id: I10e0c49b0d89b9530a523e9db79d5c7aa52359a4