Remove tripleo jobs from stable/ussuri branch
In TripleO, os_tempest is used from master for all
release. It does not make sense to keep running these jobs
in os_tempest for stable/ussuri branch.

This job is also based on old zuul parenting work, which are
getting removed via this topic:
https://review.opendev.org/q/topic:post_cleanup_zuul_parenting

So this patch removes these jobs completely.
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>
Change-Id: I20ecfcb95d482701e865151822c2d625c0640743