Temporarily re-add redhat-8.yml to vars
Although tripleo deprecate master, we still use os_tempest in wallaby
and train lines, and due a design error, we rely on os_tempest master
branch instead of use the appropriate branch, and the removal of
redhat-8.yml file broken these jobs.
We will in the future, use the proper os_tempest branch on these jobs.

Change-Id: Ic2c547d730e900c883685050366e8e307869e6d1