Add missed openstack_resources_image condition
In some jobs where we run tempest we do not have all required services
Good example of that are infra jobs where only keystone is deployed.
During conversion to usage of openstack_resources role, a condition for
glance images creation task was missed, that avoided image uploads
when glance is not present.

Change-Id: Ibcf16abab20eadf984ad1c8307ad14b43a889793